# apps/server/auth-service/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/social/{provider}/init": {
      "get": {
        "operationId": "AuthController_initSocialLogin",
        "parameters": [
          {
            "name": "platform",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "web",
                "mobile"
              ],
              "type": "string"
            }
          },
          {
            "name": "invite",
            "required": false,
            "in": "query",
            "description": "Invitation token: accept the invitation with Google",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Persist platform then redirect into the provider OAuth flow",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/oidc/login": {
      "get": {
        "operationId": "AuthController_oidcLogin",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Begin OIDC SSO (redirects to the IdP)",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/exchange": {
      "post": {
        "operationId": "AuthController_exchange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MFA_REQUIRED (Google sign-in of an Owner / Admin-like role, or of a member who turned 2FA on), else tokens (incl. OIDC SSO). 403 SSO_REQUIRED when the workspace requires SSO for this member",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ExchangeResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/MfaRequiredResponseDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Exchange a one-time login code for tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New access token issued"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Rotate the access token using a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "MFA_REQUIRED (2FA step: always for Owner / Admin-like roles, for other members once they turned 2FA on), else LOGIN_SUCCESS + tokens",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/LoginTokensResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/MfaRequiredResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "SSO_REQUIRED: the workspace requires SSO for this email domain"
          }
        },
        "summary": "Authenticate with email + password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke the current session",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgot",
        "parameters": [
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Send a password-reset OTP to the email",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Verify a password-reset OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/resend-otp": {
      "post": {
        "operationId": "AuthController_resend",
        "parameters": [
          {
            "name": "accept-language",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Resend a password-reset OTP",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reset password using a verified OTP",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Platform Auth Service",
    "description": "Authentication, OTP, refresh tokens, user search and friends API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ExchangeUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "isVerified": {
            "type": "boolean"
          },
          "mustSetPassword": {
            "type": "boolean",
            "description": "true only for an account created by accepting an invitation with Google: the client must show \"create your PON password\" before the app"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "isVerified",
          "mustSetPassword"
        ]
      },
      "ExchangeResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "sid": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/ExchangeUserDto"
          }
        },
        "required": [
          "userId",
          "sid",
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginTokensUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "mustSetPassword": {
            "type": "boolean",
            "description": "true only for an account created by accepting an invitation with Google: the client must show \"create your PON password\" before the app"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "mustSetPassword"
        ]
      },
      "LoginTokensResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "LOGIN_SUCCESS",
            "description": "LOGIN_SUCCESS (login, MFA verify / enroll-complete, accept-password); INVITATION_ACCEPTED is no longer emitted"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "sid": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/LoginTokensUserDto"
          }
        },
        "required": [
          "code",
          "accessToken",
          "refreshToken",
          "sid",
          "user"
        ]
      },
      "MfaUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "displayName"
        ]
      },
      "MfaRequiredResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "MFA_REQUIRED"
            ],
            "example": "MFA_REQUIRED"
          },
          "mfaToken": {
            "type": "string",
            "description": "Opaque, single-use, valid for 5 minutes"
          },
          "enrollmentRequired": {
            "type": "boolean",
            "description": "true = 2FA is mandatory for the role (Owner / Admin-like) and the account never enrolled: set up an authenticator first. false = enter a code (enrolled; for a Member, 2FA they turned on)"
          },
          "user": {
            "$ref": "#/components/schemas/MfaUserDto"
          }
        },
        "required": [
          "code",
          "mfaToken",
          "enrollmentRequired",
          "user"
        ]
      },
...
      "ExchangeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "One-time login code returned by the OAuth/login flow"
          },
          "deviceId": {
            "type": "string",
            "description": "Device identifier for the issued session"
          },
          "platform": {
            "type": "string",
            "description": "Originating platform",
            "example": "web"
          }
        },
        "required": [
          "code"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "sid": {
            "type": "string",
            "description": "Session id issued at login"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token issued at login"
          }
        },
        "required": [
          "sid",
          "refreshToken"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Registered email address"
          },
          "password": {
            "type": "string",
            "example": "P@ssw0rd123",
            "description": "Account password",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email to send the reset OTP to"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email the OTP was sent to"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "One-time code from the email"
          }
        },
        "required": [
          "email",
          "otp"
        ]
      },
      "ResendOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email to resend the OTP to"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email of the account to reset"
          },
          "otp": {
            "type": "string",
            "example": "123456",
            "description": "Verified OTP code"
          },
          "password": {
            "type": "string",
            "example": "N3wP@ssw0rd",
            "description": "New password",
            "minLength": 8
          }
        },
        "required": [
          "email",
          "otp",
          "password"
        ]
      },
...
    }
  }
}